TechSprint/Prototipos/PS1 - Team 1/requirements.txt
streamlit
pandas
requests
langchain
openai
python-dotenv
langchain-openai
langchain-community
tabulate
fastapi
uvicorn
pydantic
gdown
google-cloud-translate
